tests/wpt/web-platform-tests/offscreen-canvas/text/2d.text.font.relative_size.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.text.font.relative_size</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>2d.text.font.relative_size</h1>
<p class="desc"></p>


<script>
var t = async_test("");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

ctx.font = '1em sans-serif';
_assertSame(ctx.font, '10px sans-serif', "ctx.font", "'10px sans-serif'");

t.done();

});
</script>
